refactor: move userIsDeveloperOfAppWithUuid to own file
